requirements.txt
pandas
faker
sqlalchemy
psycopg2-binary
pytest
